# Improve the performance of the `ANY` aggregation function by acting over batches.