Changelog category (leave one):
- Performance Improvement


Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Improve the performance of the `ANY` aggregation function by acting over batches.


> Information about CI checks: https://clickhouse.tech/docs/en/development/continuous-integration/

Stop processing data as soon as an element has been saved. In most situations, i.e. when some element is saved, this leads to nice improvements:

Quick benchmark:

### Any
```
./programs/clickhouse benchmark --query 'Select any(number) from numbers(1000000000)' -i 20
```

* Master


```
Queries executed: 20.
localhost:9000, queries 20, QPS: 3.481, RPS: 3480741146.794, MiB/s: 26555.947, result RPS: 3.481, result MiB/s: 0.000.
0.000%          0.283 sec.
```

* Changes:
```
Queries executed: 20.
localhost:9000, queries 20, QPS: 3.845, RPS: 3845155687.009, MiB/s: 29336.210, result RPS: 3.845, result MiB/s: 0.000.
0.000%          0.252 sec.
```


### AnyIf

```
./programs/clickhouse benchmark --query 'Select anyIf(number, number > 0) from numbers(1000000000)' -i 20
```

* Master
```
Queries executed: 20.
localhost:9000, queries 20, QPS: 0.679, RPS: 679409111.125, MiB/s: 5183.480, result RPS: 0.679, result MiB/s: 0.000.
0.000%          1.397 sec.
```


* Changes

```
Queries executed: 20.
localhost:9000, queries 20, QPS: 1.086, RPS: 1085583240.480, MiB/s: 8282.343, result RPS: 1.086, result MiB/s: 0.000.
0.000%          0.894 sec
```

### AnyIf (Nullable)

```
./programs/clickhouse benchmark --query 'Select anyIf(number, number > 0) from (Select toNullable(number) as number FROM numbers(1000000000))' -i 20
```

* Master

```
Queries executed: 20.
localhost:9000, queries 20, QPS: 0.673, RPS: 672793289.616, MiB/s: 5133.005, result RPS: 0.673, result MiB/s: 0.000.
0.000%          1.414 sec.

```


* Changes:

```
Queries executed: 20.
localhost:9000, queries 20, QPS: 0.936, RPS: 936020885.318, MiB/s: 7141.273, result RPS: 0.936, result MiB/s: 0.000.
0.000%          1.029 sec.
```


Note that this only applies to ANY(). I might end up specialising things even more in further PRs to improve other aggregate functions.